Conclusion, Caption, Icon Trouble
Added conclusion, caption for photo and playing around with markers a bit. I'll add some more with the styling tonight.

Can you check the formatting of the caption? Is it the right size/style?

Also, I'm trying to play around with the icon, and on line 21 when I change it to "L.icon({..." it blows up and it wont let me play with all of the options for icons from the Leaflet documentation. Any ideas on whats going on? @thomaswilburn