Key
One more thing. I'm trying to get a map key in here (beta test with my dad lead to the discovery of needing a key). I wrote it in to the story.txt, but i'm a little confused on where to put it in the HTML. Ideally, I was thinking that the key would be a fixed box on the bottom right of the map. Any ideas on how to get it to pop in? @thomaswilburn