Added in Politics Text, A few more ideas
I added in the key after the introduction, do you think this is effective? Also, do you think it'd make sense to add in a second highlight color (red?) for cities that don't meet crtieria? Like for congestion, potentially it would make more sense to highlight the worst cities in the country with a different color to differentiate the types of highlight? If so, what is the best way to do that? @thomaswilburn